+++ packages/blocks-helpers/src/default-attributes/style.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import type { DefaultAttributes } from "@atomicsmash/blocks-helpers";
 
 const validBackgroundValue = {